Update mute query selector for new UI.
Google is rolling out a new Meet UI. The mute div has become a button.

See https://cloud.google.com/blog/products/google-meet/new-features-for-google-meet

This change is meant to work on both the old and new version as they are
both still live.

Tested manually on both versions of Meet.