Click the mute button rather than using key shortcut
It appears as though key shortcuts might vary across
platforms. The previous method of toggling mute was
to press META+d, which is Command+d on MacOS, but ends
up being WindowsKey+d on Windows or Linux PCs for the
most part, which doesn't match the CTRL+d shortcut
that meet wants when running on those systems.

Rather than trying to figure out the underlying system
to guess the correct keybindings, or just risking more
issues later if Google decides to update they key shortcut,
just have the extension click the mute button, since it
already has a selector for it that is required for it to
function correctly.